Made the 'AudioResourceID' failure as 0 instead of -1

+++ DeadFrame2D/include/Data/Systems/Audio/AudioResourceID.h
@@ -4,5 +4,5 @@
 
 namespace DF2D::Data
 {
-	using AudioResourceID = int64_t;
+	using AudioResourceID = uint32_t;
 }
